stock 타입 Long -> Integer로 변경

+++ module-shop/src/main/java/Project/moduleshop/entity/Product.java
@@ -19,7 +19,7 @@ public class Product {
     private Long id;
     private String name;
     private Long price;
-    private Long stock;
+    private Integer stock;
     private String explain;
     private LocalDateTime openAt;
 }
